+++ src/es6/docs.js
@@ -1,7 +1,5 @@
 (() => {
-  const demo = [1, 2, 3].map(v => v * v)
-  console.log(demo.toString())
+  // coding here
 
-  const [pm, cm] = [200, 400]
-  console.log(pm, cm)
+  console.log('hello!')
 })()